creating-video/requirements.txt
pyyaml
numpy
imageio
imageio[ffmpeg]
pytest
